Here we go again

+++ test/t8_cmesh/t8_gtest_compute_first_element.cxx
@@ -172,7 +172,7 @@ TEST_P (t8_gtest_rank_times_global_num_elems_over_size, large_numbers)
         const uint64_t sum_3 = (remainder_0 * remainder_1) / mpisize_64;
 
         const uint64_t computed_result = sum_0 + sum_1 + sum_2 + sum_3;
-        check_result = (rank == size) ? num_elems : check_result;
+        check_result = (((uint64_t) rank) == size) ? num_elems : check_result;
         // if(rank==size) {
         //   t8_productionf("ranke==size\n");
         //   check_result = num_elems;
